Added package recreating nav2 tutorial up to odometry.
The "sam bot" is replaced with a preliminary model of our forklift. To be updated. Many errors where encountered and fixed while following the tutorial, so it is recommended to install the "forklift_bot_description" package (found in EXPERIMENTATION). This is done as usual by cd into the forklift_bot_description folder and then "colcon build", ". install/setup.bash", finally launch with "ros2 launch forklift_bot_description display.launch.py"